Улучшена структуризация проекта, добавлена оптимизация для API методов
