+++ tmsdp/tms/utils/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from . import models
+from ..database import models
 
 
 class TestProjectCreateSerializer(serializers.ModelSerializer):
